+++ server/helpers/data.js
@@ -193,16 +193,16 @@ module.exports = {
       NATURAL JOIN
         ${hydrusTables.tags}
       NATURAL JOIN
-        ${hydrusTables.repositoryHashIdMap}
+        ${hydrusTables.repositoryHashIdMapTags}
       NATURAL JOIN
         ${hydrusTables.filesInfo}
       WHERE
-        ${hydrusTables.tags}.tag LIKE ?
+        ${hydrusTables.tags}.tag LIKE '%:%'
       AND
-        ${hydrusTables.filesInfo}.mime IN (${hydrusConfig.mimePlaceholders})
+        ${hydrusTables.filesInfo}.mime IN (${hydrusConfig.supportedMimeTypes})
       ORDER BY
         name`
-    ).pluck().all(`%:%`, hydrusConfig.supportedMimeTypes)
+    ).pluck().all()
   },
   getTags () {
     return db.hydrus.prepare(
@@ -216,33 +216,153 @@ module.exports = {
       NATURAL JOIN
         ${hydrusTables.tags}
       NATURAL JOIN
-        ${hydrusTables.repositoryHashIdMap}
+        ${hydrusTables.repositoryHashIdMapTags}
       NATURAL JOIN
         ${hydrusTables.filesInfo}
       WHERE
-        ${hydrusTables.filesInfo}.mime IN (${hydrusConfig.mimePlaceholders})`
-    ).all(hydrusConfig.supportedMimeTypes)
+        ${hydrusTables.filesInfo}.mime IN (${hydrusConfig.supportedMimeTypes})`
+    ).all()
   },
   getFiles (namespaces) {
+    const serviceHashIds = db.hydrus.prepare(
+      `SELECT
+        ${hydrusTables.currentFiles}.service_hash_id AS id
+      FROM
+        ${hydrusTables.currentFiles}`
+    ).all().map(row => row.id)
+
+    let possibleRowsA = db.hydrus.prepare(
+      `SELECT
+        ${hydrusTables.repositoryHashIdMapFiles}.service_hash_id
+          AS serviceHashId,
+        ${hydrusTables.repositoryHashIdMapFiles}.master_hash_id
+          AS masterHashId,
+        ${hydrusTables.repositoryHashIdMapFiles}.hash_id_timestamp
+          AS timestamp
+      FROM
+        ${hydrusTables.repositoryHashIdMapFiles}
+      WHERE
+        ${hydrusTables.repositoryHashIdMapFiles}.service_hash_id
+          IN (${serviceHashIds})`
+    ).all()
+
+    let possibleRowsB = db.hydrus.prepare(
+      `SELECT
+        ${hydrusTables.repositoryHashIdMapTags}.service_hash_id
+          AS serviceHashId,
+        ${hydrusTables.repositoryHashIdMapTags}.master_hash_id
+          AS masterHashId,
+        ${hydrusTables.repositoryHashIdMapTags}.hash_id_timestamp
+          AS timestamp
+      FROM
+        ${hydrusTables.repositoryHashIdMapTags}
+      WHERE
+        ${hydrusTables.repositoryHashIdMapTags}.service_hash_id
+          IN (${serviceHashIds})`
+    ).all()
+
+    let finalRows = []
+
+    let usedServiceHashIds = []
+    let usedMasterHashIds = []
+
+    for (let i = 0; i < possibleRowsA.length; i++) {
+      const a = possibleRowsA[i]
+
+      for (let j = 0; j < possibleRowsB.length; j++) {
+        const b = possibleRowsB[j]
+
+        if (
+          a.serviceHashId === b.serviceHashId &&
+          (!usedServiceHashIds.includes(a.serviceHashId))
+        ) {
+          if (a.timestamp > b.timestamp) {
+            finalRows.push(a)
+          } else {
+            finalRows.push(b)
+          }
+
+          possibleRowsA.splice(i, 1)
+          possibleRowsB.splice(j, 1)
+
+          usedServiceHashIds.push(a.serviceHashId)
+          usedServiceHashIds.push(b.serviceHashId)
+          usedMasterHashIds.push(a.masterHashId)
+          usedMasterHashIds.push(b.masterHashId)
+        }
+
+        if (
+          a.masterHashId === b.masterHashId &&
+          (!usedMasterHashIds.includes(a.masterHashId))
+        ) {
+          if (a.timestamp > b.timestamp) {
+            finalRows.push(a)
+          } else {
+            finalRows.push(b)
+          }
+
+          possibleRowsA.splice(i, 1)
+          possibleRowsB.splice(j, 1)
+
+          usedServiceHashIds.push(a.serviceHashId)
+          usedServiceHashIds.push(b.serviceHashId)
+          usedMasterHashIds.push(a.masterHashId)
+          usedMasterHashIds.push(b.masterHashId)
+        }
+      }
+    }
+
+    usedServiceHashIds = [...new Set(usedServiceHashIds)]
+    usedMasterHashIds = [...new Set(usedMasterHashIds)]
+
+    possibleRowsA = possibleRowsA.filter(
+      row => (!usedServiceHashIds.includes(row.serviceHashId))
+    )
+    possibleRowsA = possibleRowsA.filter(
+      row => (!usedMasterHashIds.includes(row.masterHashId))
+    )
+
+    possibleRowsB = possibleRowsA.filter(
+      row => (!usedServiceHashIds.includes(row.serviceHashId))
+    )
+    possibleRowsB = possibleRowsA.filter(
+      row => (!usedMasterHashIds.includes(row.masterHashId))
+    )
+
+    finalRows = finalRows.concat(possibleRowsA)
+    finalRows = finalRows.concat(possibleRowsB)
+
+    const mappedServiceHashIds = []
+
+    for (const finalRow of finalRows) {
+      mappedServiceHashIds[finalRow.masterHashId] = finalRow.serviceHashId
+    }
+
+    const masterHashIds = finalRows.map(
+      row => row.masterHashId
+    )
+
     const files = db.hydrus.prepare(
       `SELECT
-        ${hydrusTables.currentFiles}.service_hash_id AS id,
+        ${hydrusTables.hashes}.master_hash_id AS masterHashId,
+        ${hydrusTables.hashes}.hash,
         ${hydrusTables.filesInfo}.mime AS mimeType,
         ${hydrusTables.filesInfo}.size,
         ${hydrusTables.filesInfo}.width,
-        ${hydrusTables.filesInfo}.height,
-        ${hydrusTables.hashes}.hash
+        ${hydrusTables.filesInfo}.height
       FROM
-        ${hydrusTables.currentFiles}
-      NATURAL JOIN
-        ${hydrusTables.repositoryHashIdMap}
-      NATURAL JOIN
         ${hydrusTables.hashes}
       NATURAL JOIN
         ${hydrusTables.filesInfo}
       WHERE
-        ${hydrusTables.filesInfo}.mime IN (${hydrusConfig.mimePlaceholders})`
-    ).all(hydrusConfig.supportedMimeTypes)
+        ${hydrusTables.hashes}.master_hash_id IN (${masterHashIds})
+      AND
+        ${hydrusTables.filesInfo}.mime IN (${hydrusConfig.supportedMimeTypes})`
+    ).all()
+
+    for (const file of files) {
+      file.id = mappedServiceHashIds[file.masterHashId]
+    }
 
     for (const namespace of namespaces) {
       for (const file of files) {
@@ -256,16 +376,16 @@ module.exports = {
           NATURAL JOIN
             ${hydrusTables.tags}
           NATURAL JOIN
-            ${hydrusTables.repositoryHashIdMap}
+            ${hydrusTables.repositoryHashIdMapTags}
           WHERE
-            ${hydrusTables.tags}.tag LIKE ?
+            ${hydrusTables.tags}.tag LIKE '${namespace}:%'
           AND
-            ${hydrusTables.repositoryHashIdMap}.service_hash_id = ?
+            ${hydrusTables.repositoryHashIdMapTags}.service_hash_id = ${file.id}
           ORDER BY
             ${hydrusTables.tags}.tag
           LIMIT
             1`
-        ).all(`${namespace}:%`, file.id)
+        ).all()
 
         file[`namespace_${namespace.split(' ').join('_')}`] = tags.length
           ? tags[0].tag.replace(`${namespace.split(' ').join('_')}:`, '')
@@ -287,11 +407,11 @@ module.exports = {
         ON ${hydrusTables.currentFiles}.service_hash_id =
           ${hydrusTables.currentMappings}.service_hash_id
       NATURAL JOIN
-        ${hydrusTables.repositoryHashIdMap}
+        ${hydrusTables.repositoryHashIdMapTags}
       NATURAL JOIN
         ${hydrusTables.filesInfo}
       WHERE
-        ${hydrusTables.filesInfo}.mime IN (${hydrusConfig.mimePlaceholders})`
-    ).all(hydrusConfig.supportedMimeTypes)
+        ${hydrusTables.filesInfo}.mime IN (${hydrusConfig.supportedMimeTypes})`
+    ).all()
   }
 }