+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.3.0] - 2018-07-15
+
 ### Added
 
 + Added check if hydrus server has already set up its repositories when
@@ -183,7 +185,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 + Initial release
 
-[Unreleased]: https://github.com/mserajnik/hydrusrv/compare/2.2.0...develop
+[Unreleased]: https://github.com/mserajnik/hydrusrv/compare/2.3.0...develop
+[2.3.0]: https://github.com/mserajnik/hydrusrv/compare/2.2.0...2.3.0
 [2.2.0]: https://github.com/mserajnik/hydrusrv/compare/2.1.0...2.2.0
 [2.1.0]: https://github.com/mserajnik/hydrusrv/compare/2.0.1...2.1.0
 [2.0.1]: https://github.com/mserajnik/hydrusrv/compare/2.0.0...2.0.1